Mostly documented prediction module
