+++ 05_prediction/src/Predictors/Predictor.py
@@ -1,0 +1,2 @@
+class Predictor:
+    pass
