+++ Pipfile
@@ -14,7 +14,7 @@ opencv-python = ">=4.4"
 tqdm = "*"
 keras = "~=2.3.1"
 segmentation-models = "~=1.0.1"
+typing-extensions = "*"
 
 [requires]
 python_version = "3.7"
-
